random seed
It seems that people were using lots of different strategies for dealing with the random seed, so to make grading easier, I've provided a little bit of skeleton code that A) should be consistent with the unit tests B) should reduce confusion C) give folks a little more guidance about how to use the code.  Check out the Github repo for the update.