Fix lda unit test
The test `test_m` attempts to test the m_step of the algorithm is working properly. However, the unit test
goes about checking that `vb._beta` is `almostEqual` to certain values, instead of checking that the return
value of `m_step` conforms to these `almostEqual` calls. Since the call to `m_step` does not actually update
`vb._beta`, it does not make sense to test that variable. This commit fixes the test by instead testing what
is returned from `m_step`.